2025.6.15 数据结构调整

+++ Transoceanic/MathHelp/BitOperation.cs
@@ -2,7 +2,7 @@
 
 public static class BitOperation
 {
-    public static bool GetBit(sbyte number, byte bitPosition)
+    public static bool GetBit(sbyte number, int bitPosition)
     {
         if (bitPosition >= 8)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -10,7 +10,7 @@ public static class BitOperation
         return (number & (1 << bitPosition)) != 0;
     }
 
-    public static bool GetBit(byte number, byte bitPosition)
+    public static bool GetBit(byte number, int bitPosition)
     {
         if (bitPosition >= 8)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -18,7 +18,7 @@ public static class BitOperation
         return (number & (1 << bitPosition)) != 0;
     }
 
-    public static bool GetBit(short number, byte bitPosition)
+    public static bool GetBit(short number, int bitPosition)
     {
         if (bitPosition >= 16)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -26,7 +26,7 @@ public static class BitOperation
         return (number & (1 << bitPosition)) != 0;
     }
 
-    public static bool GetBit(ushort number, byte bitPosition)
+    public static bool GetBit(ushort number, int bitPosition)
     {
         if (bitPosition >= 16)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -34,7 +34,7 @@ public static class BitOperation
         return (number & (1 << bitPosition)) != 0;
     }
 
-    public static bool GetBit(int number, byte bitPosition)
+    public static bool GetBit(int number, int bitPosition)
     {
         if (bitPosition >= 32)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -42,7 +42,7 @@ public static class BitOperation
         return (number & (1 << bitPosition)) != 0;
     }
 
-    public static bool GetBit(uint number, byte bitPosition)
+    public static bool GetBit(uint number, int bitPosition)
     {
         if (bitPosition >= 32)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -50,7 +50,7 @@ public static class BitOperation
         return (number & (1 << bitPosition)) != 0;
     }
 
-    public static bool GetBit(long number, byte bitPosition)
+    public static bool GetBit(long number, int bitPosition)
     {
         if (bitPosition >= 64)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -58,7 +58,7 @@ public static class BitOperation
         return (number & (1 << bitPosition)) != 0;
     }
 
-    public static bool GetBit(ulong number, byte bitPosition)
+    public static bool GetBit(ulong number, int bitPosition)
     {
         if (bitPosition >= 64)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -66,7 +66,7 @@ public static class BitOperation
         return (number & (1ul << bitPosition)) != 0;
     }
 
-    public static bool GetBit(Int128 number, byte bitPosition)
+    public static bool GetBit(Int128 number, int bitPosition)
     {
         if (bitPosition >= 128)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -74,7 +74,7 @@ public static class BitOperation
         return (number & (1 << bitPosition)) != 0;
     }
 
-    public static bool GetBit(UInt128 number, byte bitPosition)
+    public static bool GetBit(UInt128 number, int bitPosition)
     {
         if (bitPosition >= 128)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -82,7 +82,7 @@ public static class BitOperation
         return (number & (UInt128.One << bitPosition)) != 0;
     }
 
-    public static sbyte SetBit(sbyte number, byte bitPosition, bool value)
+    public static sbyte SetBit(sbyte number, int bitPosition, bool value)
     {
         if (bitPosition >= 8)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -90,7 +90,7 @@ public static class BitOperation
         return (sbyte)((number & ~(1 << bitPosition)) | (value.ToInt() << bitPosition));
     }
 
-    public static byte SetBit(byte number, byte bitPosition, bool value)
+    public static byte SetBit(byte number, int bitPosition, bool value)
     {
         if (bitPosition >= 8)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -98,7 +98,7 @@ public static class BitOperation
         return (byte)((number & ~(1 << bitPosition)) | (value.ToInt() << bitPosition));
     }
 
-    public static short SetBit(short number, byte bitPosition, bool value)
+    public static short SetBit(short number, int bitPosition, bool value)
     {
         if (bitPosition >= 16)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -106,7 +106,7 @@ public static class BitOperation
         return (short)((number & ~(1 << bitPosition)) | (value.ToInt() << bitPosition));
     }
 
-    public static ushort SetBit(ushort number, byte bitPosition, bool value)
+    public static ushort SetBit(ushort number, int bitPosition, bool value)
     {
         if (bitPosition >= 16)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -114,7 +114,7 @@ public static class BitOperation
         return (ushort)((number & ~(1 << bitPosition)) | (value.ToInt() << bitPosition));
     }
 
-    public static int SetBit(int number, byte bitPosition, bool value)
+    public static int SetBit(int number, int bitPosition, bool value)
     {
         if (bitPosition >= 32)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -122,7 +122,7 @@ public static class BitOperation
         return (number & ~(1 << bitPosition)) | (value.ToInt() << bitPosition);
     }
 
-    public static uint SetBit(uint number, byte bitPosition, bool value)
+    public static uint SetBit(uint number, int bitPosition, bool value)
     {
         if (bitPosition >= 32)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -130,7 +130,7 @@ public static class BitOperation
         return (number & ~(1u << bitPosition)) | ((uint)value.ToInt() << bitPosition);
     }
 
-    public static long SetBit(long number, byte bitPosition, bool value)
+    public static long SetBit(long number, int bitPosition, bool value)
     {
         if (bitPosition >= 64)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -138,7 +138,7 @@ public static class BitOperation
         return (number & ~(1 << bitPosition)) | ((long)value.ToInt() << bitPosition);
     }
 
-    public static ulong SetBit(ulong number, byte bitPosition, bool value)
+    public static ulong SetBit(ulong number, int bitPosition, bool value)
     {
         if (bitPosition >= 64)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -146,7 +146,7 @@ public static class BitOperation
         return (number & ~(1ul << bitPosition)) | ((ulong)value.ToInt() << bitPosition);
     }
 
-    public static Int128 SetBit(Int128 number, byte bitPosition, bool value)
+    public static Int128 SetBit(Int128 number, int bitPosition, bool value)
     {
         if (bitPosition >= 128)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
@@ -154,7 +154,7 @@ public static class BitOperation
         return (number & ~(1 << bitPosition)) | (value.ToInt() << bitPosition);
     }
 
-    public static UInt128 SetBit(UInt128 number, byte bitPosition, bool value)
+    public static UInt128 SetBit(UInt128 number, int bitPosition, bool value)
     {
         if (bitPosition >= 128)
             throw new ArgumentOutOfRangeException(nameof(bitPosition));
